# field-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.TimeRequest:
    properties:
      end_time:
        type: string
      start_time:
        type: string
    required:
    - end_time
    - start_time
    type: object
  dto.TimeResponse:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message: {}
      status:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A microservice for managing booking system fields, operational times,
    and schedules.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Field Service API
  version: "1.0"
paths:
...
  /time:
    get:
      consumes:
      - application/json
      description: Get a list of all operational times
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TimeResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all operational times
      tags:
      - times
    post:
      consumes:
      - application/json
      description: Create a new operational time slot
      parameters:
      - description: Create Time Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TimeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TimeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new operational time
      tags:
      - times
  /time/{uuid}:
    get:
      consumes:
      - application/json
      description: Get detailed information about an operational time
      parameters:
      - description: Time UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TimeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get time detail by UUID
      tags:
      - times
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"